requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rapidfuzz>=3.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.28.0
scikit-learn>=1.3.0
